+++ Projects/Payment/GraphQL/Octopus/PaymentInformation.graphql
@@ -7,5 +7,11 @@ query PaymentInformation {
       }
       status
     }
+    upcomingCharge {
+      date
+    }
+    chargeHistory {
+      status
+    }
   }
 }
